LayoutTests/svg/dynamic-updates/SVGFESpotLightElement-dom-pointsAtY-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'pointsAtY' attribute of the SVGFESpotLightElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS spotLightElement.getAttribute('pointsAtY') is "0"
PASS spotLightElement.getAttribute('pointsAtY') is "100"
PASS successfullyParsed is true

TEST COMPLETE
